home work 4

+++ Web_lesson4/sort.py
@@ -10,23 +10,19 @@ def handle_file(file: Path, root_folder: Path, dist):
     target_folder = root_folder / dist
     target_folder.mkdir(exist_ok=True)
     ext = Path(file).suffix
-    new_name = normalize(file.name.replace(ext, "")) + ext
-    file.replace(target_folder / new_name)
-
-def handle_archive(file: Path, root_folder: Path, dist):
-    target_folder = root_folder / dist
-    target_folder.mkdir(exist_ok=True)  # create folder ARCH
-    ext = Path(file).suffix
-    folder_for_arch = normalize(file.name.replace(ext, ""))
-    archive_folder = target_folder / folder_for_arch
-    archive_folder.mkdir(exist_ok=True)  # create folder ARCH/name_archives
-    try:
-        shutil.unpack_archive(str(file.resolve()), str(archive_folder.resolve()))
-    except shutil.ReadError:
-        archive_folder.rmdir()  # Если не успешно удаляем папку под  архив
-        return
-    file.unlink()  # Если успешно удаляем архив
-
+    if dict == "ARCH":
+        folder_for_arch = normalize(file.name.replace(ext, ""))
+        archive_folder = target_folder / folder_for_arch
+        archive_folder.mkdir(exist_ok=True)  # create folder ARCH/name_archives
+        try:
+            shutil.unpack_archive(str(file.resolve()), str(archive_folder.resolve()))
+        except shutil.ReadError:
+            archive_folder.rmdir()
+            return
+        file.unlink()
+    else:
+        new_name = normalize(file.name.replace(ext, "")) + ext
+        file.replace(target_folder / new_name)
 
 def handle_folder(folder: Path):
     try:
@@ -74,7 +70,7 @@ def main(folder):
             executor.submit(handle_file, file, folder, "OTHER")
 
         for file in scan.ARCH:
-            executor.submit(handle_archive, file, folder, "ARCH")
+            executor.submit(handle_file, file, folder, "ARCH")
 
     print('End')
     for f in scan.FOLDERS:
